Rename fpu_driver to runtime
